Take g = middle of double-precision interval
This computes a double precision interval for possible g values. Instead
of finding the simplest rational value in that interval, this just uses
the midpoint of the interval as the value of g. This guarantees hitting
the end-points exactly and the zero point is inherently hit since all
range values are multiples of g and the zero point is 0g. It's unclear
if this works in all cases, but it's a nice simple approach.