fix g_ival in case where h⁻ ≠ h or h⁺ ≠ h after canonicalization
This issue here was that we were computing l⁻/l⁺ as the error from h to
get to the halfway point between floats and then canonicalize, but that
can lead to the l⁻/l⁺ values being very far away from the desired value
so the while loops ran for a long time. This changes the approach to use
h to compute h⁻/h⁺ first and then computing l⁻/l⁺ based on that. This
guarantees that the pair is already canonicalized so we're always close
to the final value we want.