Simple example where midpoint approach fails: 0.2:0.1:1.1
    julia> lift_range(0.2, 0.1, 1.1) |> collect
    10-element Vector{Float64}:
    0.2
    0.30000000000000004
    0.4
    0.5
    0.6000000000000001
    0.7000000000000001
    0.8
    0.9
    1.0
    1.1

The problem here is that a and b are both higher than the true values so
the midpoint of feasible values is too big as well and we end up with
internal range values that are too big---specifically at those points
where the true value would be on the smaller side like 0.3. This example
shows that we do need to find the simplest rational in the g interval.