g_ival: search both ways to hit the critical point
This new example was causing a failing assertion in g_ival, which this
fixes by not assuming that a single nextfloat/prevfloat gets you in the
right region. I think this is only an issue with large c/d/e values.
Instead this uses arithmetic to get in roughly the right spot and then
two while loops in different directions to find the critical point.

It this turns out to take too long in some cases it could be changed to
a widening binary search, but that really seems like overkill.