Pick simplest d that we can
My proof seems to indicate that we want the smallest d, but there seem
to be cases where that doesn't quite do what we want. But we're using
the shortcut of looking for f⁻/2^p ≤ c/d ≤ f⁺/2^p which may behave a bit
differently. In any case, choosing the simplest in terms of powers of 2
seems to do the trick here.