All tests pass: fixed logic for finding d
We need to simultaneously find minimal d such that both of these hold:

     d*l_a ≤ c′*2^p ≤ d*h_a
     d*l_b ≤ e′*2^p ≤ d*h_b

Fortunately, n = n′*2^p is also divisible by 2^p, so we can pick either
side and find the simplest rational in it and then find our d value. We
Use ratio with the smaller absolute value and thus greater precision.